lib: tst_kconfig: Fix segfaults on variable value mismatch
If we add CONFIG_FOO=value to the test .needs_kconfig while the kernel
config will not define the particular value we will blindly call
strlen() on the val->val and end up with SegFault. Fix this by checking
that the variable has value set first.

Also initialize the val->val to NULL so that we get segfaults 100% of
the time when we attempt to dereference the value when not defined.

+ Test

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>