hugeshmat04: try to find unmapped range for test
The test intermittently fails on ppc64le, when heap
happens to overlap with segment that the test is trying
to attach at 1GB boundary.

Try to find unmapped range. In my testing, address above 1GB appears
to be needed for reproducer to work. Also add kernel commit that fixed
it to metadata.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>