controllers/pids.sh: get available for cgroup v2
controllers/pids/pids.sh doesn't work in the cgroup v2 environment because
some interfaces of cgroup v2 are different from cgroup v1.

The test itself is useful for cgroup v2 as well, so let's get it
available for cgroup v2.

The test starts with cgroup v2 if it's mounted already. On some systems,
like as Fedora 35 and CentOS Stream 9, systemd mounts cgroup v2 as the
default while system booting.

Signed-off-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>