lib: Add .skip_in_compat flag
Some tests can never be run under 32-bit compatibility mode. This adds
a flag to skip them. This will show up in the meta data in addition to
causing the test to exit with TCONF if compat mode is detected at
runtime.

It's possible that support for compat mode will be added to
CAN. If this happens then we will have to probe the interface for
support when in 32-bit compat mode.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>