posix/pthread_create/15-1: Supress warning
generate-makefiles.sh gave the following warning:
conformance/interfaces/pthread_create/15-1.c should be test.

The script has a strange implementation, that just greps for "main" in the c file.
Based on that it decides if the file should be a test or a run-test.
Other tests in the suite including the main-function from testfrmw/threads_scenarii.c
also use a comment, that supresses the message

Signed-off-by: Joerg Vehlow <joerg.vehlow@aox-tech.de>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>