posix/conformance/interfaces: Fix unsued-variable for testfrmw
Thre is a global variable sc in testfrmw/threads_scenarii, that is used,
if the integrated main function (STD_MAIN) is used. However some tests
did not use this integrated main, but still used the variable. To be able to
get rid of all warnings, the variable is duplicated into these tests and only
defined by threads_scenarii, if STD_MAIN is defined

Signed-off-by: Joerg Vehlow <joerg.vehlow@aox-tech.de>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>