Refactoring dio_sparse.c using LTP API
Note that the -a (alignment) parameter was removed from the test. The
reason is that the value provided by that argument is dumbious and even
if it did have any influence on the end result we do not have the
alignment under controll anyway. We do pass an alignment requirement to
the function that allocates the buffer, but nothing stops it from being
aligned by multiple of the requested alignment by an accident.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>