lib: add functions to adjust oom score
This introduces function to LTP to adjust the oom_score_adj of target
process, which may be helpful in OOM tests to prevent kernel killing
the main or lib process during test run.

The exported global tst_enable_oom_protection function can be used
at anywhere you want to protect, but please remember that if you
do enable protection on a process($PID) that all the children will
inherit its score and be ignored by OOM Killer as well. So that's
why tst_disable_oom_protection to be used in combination.

Signed-off-by: Li Wang <liwang@redhat.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>