syscalls/quotactl08: Test quoatctl01 but quota info hidden in filesystem
It uses two variants(quotactl and quotactl_fd). The difference for quotactl01
is that we don't use quotacheck command and quota info hidden in filesystem.

It also needs mkfs.ext4 supports quota feature.  This feature is disabled[1]
before e2fsprog 1.42 when not specifying a --enable-quota option in compile-time
and e2fsprog 1.43 remove this disable[2]. So we should limit the e2fsprog to 1.43.

[1]https://ext4.wiki.kernel.org/index.php/Quota
[2]https://git.kernel.org/pub/scm/fs/ext2/e2fsprogs.git/commit/?h=v1.43&id=9e8fcd6e

Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Yang Xu <xuyang2018.jy@fujitsu.com>