chdir01: Fix on more restrictive umask
Fixes the test on more restrictive umask, this was found on FIPS but
it's not limited to the FIPS system at all. All that is needed to
reproduce the problem is to set umask to 077 before the test is
executed.

The troublemaker here is the FAT filesystem since there are no access
bits defined in FAT disk format and they are emulated completely by the
kernel. This means that all files on FAT filesytem have access bits
generated by the kernel accordingly to system umask at the time the
filesystem was mounted. This means that in order to be able to access
the files on FAT we have to set the umask(0) before we mount the
filesystem.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>