syscalls/quotactl07: Refactor this case
man-pages has error that Q_XQUOTARM was not introduced by kernel
3.16[1][2]. It only doesn't have owner handler so the flag spaces overlap
a bit[3].

Add a functional test for Q_XQUOTARM and check whether available block
is greater than before instead of a regression test for buggy 3.16 kernel
because it is about 8 year old. Functional test is more meaningful.

Since kernel commit 40b52225e ("xfs: remove support for disabling quota
accounting on a mounted file system")[4] was introduced, if we want to
disable quota account feature before Q_XQUOTARM, we must need to remount
with noquota like xfs maintainer does in xfstests xfs/220 [5].

So, change this by mount with usrquota and remount with noquota to create
a test environment that superblock has quota data but quota feature not running.

We also add quotactl_fd test variant for this test.

[1]https://github.com/alejandro-colomar/man-pages/commit/38bccbcf4f51c5370a1060e6a80b90d68b0dcdc8
[2]https://github.com/alejandro-colomar/man-pages/commit/26f3978f04a1aeeb5397a5facebaef40a341afb6
[3]https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=9da93f9b7c
[4]https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit?id=40b52225e58
[5]https://patchwork.kernel.org/project/fstests/patch/20220105195352.GM656707@magnolia/

Signed-off-by: Yang Xu <xuyang2018.jy@fujitsu.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>