utime03.c: Drop 9p whitelist
Reasons:
1) It's whitelisted as "v9", which is wrong: it'd have to be "9p" (which
is in tst_fs_type_name(), but even though it would not be considered
because test is using .all_filesystems = 1 and "9p" it's not in
tst_fs_type_name() (the list of filesystems being used for .all_filesystems).

2) Tested - not needed. Working on 9p mounted on guests VMs with 5.16,
5.2.1 and 4.4.21 (on host with 5.16). The whitelist was added in 2011
(bc5da68248), i.e. kernel 2.6.

Use the same name as in tst_fs_type_name() to actually skip it.

Acked-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Li Wang <liwang@redhat.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>