Reduce default number of threads in dio_sparse.c
Initialized options out of setup and reduced default number of
threads from 1000 to 16.

Previously the default number of children was 1 so the change to 1000
made quite a difference and the test would run for half an hour on
slower hardware without any parameters. However given majority of hardware today
has more than 2 CPUs defaulting to 1 child is not that useful so the
number 16 was choosen as a good compromise.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.de>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>