ptrace08.c: Remove kernel version dependence
This test falsely fails on some ditributives because of
inappropriate kernel version comparison. For example on
VzLinux 8 it writes examining kernel address to dr0 and
doesn't check if the address was realy written. Then it
checks if enabling debug in dr7 is successful.  Because
the restricted address was not really written to dr0, dr7 writing is
successful and the result is TFAIL.

Remove kernel dependence and check if the testing address
was really written before triyng to write to dr7.

Closes: #904

Reviewed-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Alexander Ivanov <alexander.ivanov@virtuozzo.com>