memcontrol04: Copy from kselftest
At first glance this test looks the same as memcontrol03. However
there are some significant changes which complicate combining the two.

At least up to kernel (5.17) there appears to be a bug causing low
memory events in F. There is presently no solution, so a known-bug tag
with a link to Michal's investigation has been included.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Cc: Michal Koutný <mkoutny@suse.com>