tst_test.sh: Properly init $TST_FS_TYPE
to sync with C API init $TST_FS_TYPE with $LTP_DEV_FS_TYPE, when
defined. That's equivalent of the .dev_fs_type in the tst_test
structure:

    if (tst_test->dev_fs_type)
	    tdev.fs_type = tst_test->dev_fs_type;
    else
	    tdev.fs_type = tst_dev_fs_type();

(And the tst_dev_fs_type() returns either $LTP_DEV_FS_TYPE or "ext2"
if it's not defined.)

Now, when $TST_FS_TYPE properly defined, tst_mkfs() can rely on it
(thus define the check).

Also use the variable in df01.sh, mkfs01.sh (the only shell tests which
works with filesystems).

Acked-by: Yang Xu <xuyang2018.jy@fujitsu.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>