syscalls/perf_event_open03: skip test on slower systems
Some systems (specially with combination of -debug kernel
with KASAN enabled) have trouble completing this test
in specified timeout.

Lowering number of iterations would make the test condition
less accurate as it's based on global counter.

Instead, calculate the rate of iterations system can do in
first 5 seconds and used that to decide whether to continue
to run the test. If the rate is too slow, TCONF after 5
seconds.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>