getrusage03: mlock all address space for process calling consume_mb
It is trying to test without swap happen to guarantee less disturbing
for the ’->ru_maxrss‘ counting.

Therefore add mlockall() to prevent that memory address space
of the calling process from being paged to the swap area.

Due to that lock needs CAP_IPC_LOCK capability and child can not
inherit memory locks from parent, here add the definition and put
the mlockall() in consume_mb.

Signed-off-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>