ci/debian: Remove libmm-dev
It has been removed from Debian testing, thus it breaks build:
E: Unable to locate package libmm-dev
Error: Process completed with exit code 100.

But remove it for all as it was added by mistake in original Travis CI
implementation.  The only places which include mm.h are actually a
kernel modules and that include should be satisfied from kernel-devel
package.

Fixes: 24519c14a2 ("travis: Use docker based builds")

Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>