ksm06: lock memory about to be merged
We have observed sporadic failures of the test in OpenQA in the case
that the memory allocated by the test is large enough (ksm06_2 runs the
test with -n 8000 which on ppc64le with 64k page size amounts to 1GB in
allocated buffers in total) and in the case of bacground load being run
while the test is executed (LTP compilation with suitable -j make
parameter suffices to trigger the conditions).

The culprit is that some of the pages may end up being swapped before
they happen to be merged due to a memory pressure caused by the
background load.

After this patch applied the test seems to be passing regardless the
background load.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>