waitid10: Fix when coredumps are piped
When coredumps are piped into a stdin of a program (the core_pattern
starts with '|') the RLIMIT_CORE limits are not enforced.

This fixes the test expectation in the corner case that rlim_cur and
rlim_max for RLIMIT_CORE are both set to 0 but the core pattern is set
to dump core into a pipe where we previously assumed the core wouldn't
be dumped.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>